pyproject.toml
[project]
name = "context-graph"
version = "0.1.0"
description = "Traceability-first context graph for AI agents"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic>=2.9",
    "pydantic-settings>=2.5",
    "redis>=5.0",
    "neo4j>=5.25",
    "orjson>=3.10",
    "structlog>=24.4",
    "prometheus-client>=0.21",
    "litellm>=1.40",
    "openai>=1.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=6.0",
    "ruff>=0.8",
    "mypy>=1.13",
    "pre-commit>=4.0",
    "httpx>=0.28",
]
embedding = [
    "sentence-transformers>=2.6",
    "numpy>=1.26",
]
gcs = [
    "google-cloud-storage>=2.14",
]
infra-test = [
    "redis>=5.0",
    "neo4j>=5.25",
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/context_graph"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "infra: infrastructure validation tests (require docker services)",
    "unit: unit tests (no external dependencies)",
    "integration: integration tests (require docker services)",
]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "TCH"]

[tool.ruff.lint.per-file-ignores]
# Pydantic models need datetime/UUID at runtime, not just TYPE_CHECKING
"src/context_graph/domain/models.py" = ["TCH003"]

[tool.ruff.lint.isort]
known-first-party = ["context_graph"]

[tool.mypy]
python_version = "3.12"
strict = true
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = ["redis.*", "neo4j.*", "litellm", "litellm.*", "sentence_transformers", "sentence_transformers.*", "google.cloud.*", "google.auth.*", "google.api_core.*"]
ignore_missing_imports = true

[tool.coverage.run]
source = ["src/context_graph"]
omit = ["*/tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 70
